+++ DesignToolbox/DesignToolbox/Pages/Components/Dialog/Alert/AlertElements.swift
@@ -1,0 +1,42 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSSwiftUI
+import SwiftUI
+
+struct AlertElements: DesignToolboxElement {
+    let name: String
+    let illustration: AnyView
+    let pageDescription: AnyView
+
+    init() {
+        let variants: [DesignToolboxElement] = [
+            AlertMessageElement(),
+        ]
+
+        name = "app_components_alert_label".localized()
+        illustration = AnyView(AlertIllustration())
+        pageDescription = AnyView(DesignToolboxElementPage(
+            name: name,
+            illustration: illustration,
+            description: "app_components_alert_description_text",
+            demoScreen: AnyView(DesignToolboxVariantElement(elements: variants))))
+    }
+}
+
+private struct AlertIllustration: View {
+
+    var body: some View {
+        OUDSAlertMessage(label: "app_components_common_label_label".localized())
+    }
+}