+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Library] Size of loader for `button` component (Orange-OpenSource/ouds-ios#1296)
 - [Library] Size of the bullet for unordered and bare `bullet list` components (Orange-OpenSource/ouds-ios#1300)
+- [Library] Size of the bullet for unordered and bare bullet list component (Orange-OpenSource/ouds-ios#1300)
 
 ## [1.2.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.1.0...1.2.0) - 2026-02-13
 
 ### Added
 
+- [Library] `alert message` component (Orange-OpenSource/ouds-ios#1159)
 - [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
 - [Library] `bullet list` component (Orange-OpenSource/ouds-ios#513)